Add sql insertion scripts & organize database

+++ database/migrations/2018_05_15_032346_table_advertencia.php
@@ -15,12 +15,12 @@ class TableAdvertencia extends Migration
     {
         Schema::create('advertencia', function (Blueprint $table) {
             $table->increments('id');
-	    $table->tinyInteger('tipo')->unsigned();
+	    $table->Integer('tipo')->unsigned();
 	    $table->string('penalizado');
 	    $table->string('responsavel');
 	    $table->dateTime('data');
 	    $table->string('descricao');
-	    $table->tinyInteger('status')->unsigned();
+	    $table->Integer('status')->unsigned();
             $table->foreign('tipo')->references('id')->on('tipo_advertencia')->onDelete('cascade');
 	    $table->foreign('penalizado')->references('nome')->on('membro')->onDelete('cascade');
 	    $table->foreign('responsavel')->references('nome')->on('membro')->onDelete('cascade');
